+++ transformers_lightning/schedulers/cosine_scheduler_with_warmup.py
@@ -57,6 +57,6 @@ class CosineSchedulerWithWarmup(SuperScheduler):
     @staticmethod
     def add_scheduler_specific_args(parser: ArgumentParser):
         r""" Add here the hyperparameters specific of the scheduler like the number of warmup steps. """
-        super(SuperScheduler, SuperScheduler).add_scheduler_specific_args(parser)
+        super(CosineSchedulerWithWarmup, CosineSchedulerWithWarmup).add_scheduler_specific_args(parser)
         parser.add_argument('--num_warmup_steps', type=int, default=0)
         parser.add_argument('--num_cycles', type=float, default=1.0)